Revert "Introduce refactored model builder abstractions (#2241)"
This reverts commit e3365feb4729a5b32a57f1c4bbec7fff8e77c717.
