+++ src/megatron/bridge/models/model_provider.py
@@ -17,8 +17,6 @@ import os
 from pathlib import Path
 from typing import Any, Callable, Generic, TypedDict, TypeVar, Union
 
-from megatron.bridge.models.common.unimodal import _ddp_wrap, _print_num_params
-
 
 try:
     from typing import Unpack
@@ -36,7 +34,10 @@ from typing import Callable
 import torch
 from megatron.core import parallel_state, tensor_parallel
 from megatron.core.distributed import (
+    DistributedDataParallel,
     DistributedDataParallelConfig,
+    FullyShardedDataParallel,
+    TorchFullyShardedDataParallel,
 )
 from megatron.core.enums import ModelType
 from megatron.core.pipeline_parallel.utils import (
@@ -674,3 +675,84 @@ def _create_model(
             tensor_parallel.set_defaults_if_not_set_tensor_model_parallel_attributes(param)
 
     return model
+
+
+def _ddp_wrap(
+    model: list[MegatronModule],
+    data_parallel_random_init: bool,
+    ddp_config: DistributedDataParallelConfig,
+    overlap_param_gather_with_optimizer_step: bool,
+    use_megatron_fsdp: bool = False,
+    use_torch_fsdp2: bool = False,
+    *,
+    pg_collection: ProcessGroupCollection,
+) -> list[MegatronModule]:
+    """Wrap model with Distributed Data Parallel (DDP) or Fully Sharded Data Parallel (FSDP).
+
+    Args:
+        model: List of model modules to wrap
+        use_torch_fsdp2: Whether to use PyTorch FSDP v2 instead of DDP
+        data_parallel_random_init: Whether to broadcast parameters from rank 0
+        ddp_config: Configuration for distributed data parallel
+        overlap_param_gather_with_optimizer_step: Whether to disable bucketing
+            for overlapping parameter gathering with optimizer step
+
+    Returns:
+        list[MegatronModule]: List of DDP/FSDP wrapped model modules
+    """
+    if use_megatron_fsdp:
+        DP = FullyShardedDataParallel
+        if use_torch_fsdp2:
+            raise ValueError("Using use_megatron_fsdp and use_torch_fsdp2 at the same time is not supported.")
+    elif use_torch_fsdp2:
+        DP = TorchFullyShardedDataParallel
+    else:
+        DP = DistributedDataParallel
+
+    # DDP initialization is required to be on a side-stream for the full-iteration CUDA graph.
+    #  this side-stream may be nested if being called from within the get_model function, but it
+    #  is here in case someone wants to use this directly outside of get_model.
+    ddp_stream = torch.cuda.Stream()
+    ddp_stream.wait_stream(torch.cuda.current_stream())
+    with torch.cuda.stream(ddp_stream):
+        model = [
+            DP(
+                config=get_model_config(model_chunk),
+                ddp_config=ddp_config,
+                module=model_chunk,
+                # Turn off bucketing for model_chunk 2 onwards, since communication for these
+                # model chunks is overlapped with compute anyway.
+                disable_bucketing=(model_chunk_idx > 0) or overlap_param_gather_with_optimizer_step,
+                pg_collection=pg_collection,
+            )
+            for (model_chunk_idx, model_chunk) in enumerate(model)
+        ]
+    # Critical: ensure side-stream work completes before touching params on default stream
+    torch.cuda.current_stream().wait_stream(ddp_stream)
+
+    # Broadcast params from data parallel src rank to other data parallel ranks.
+    if data_parallel_random_init:
+        for model_module in model:
+            model_module.broadcast_params()
+
+    return model
+
+
+def _print_num_params(model: list[MegatronModule], pg_collection: ProcessGroupCollection) -> None:
+    """Print the number of parameters in the model on rank 0.
+
+    Only prints on data parallel rank 0 to avoid duplicate output.
+    Shows parameter count per (tensor parallel, pipeline parallel) rank.
+
+    Args:
+        model: List of model modules to count parameters from
+    """
+    if (pg_collection.dp.rank() == 0) and (pg_collection.cp.rank() == 0):
+        print(
+            " > number of parameters on (tensor, pipeline) model parallel rank ({}, {}): {}".format(
+                pg_collection.tp.rank(),
+                pg_collection.pp.rank(),
+                sum([sum([p.nelement() for p in model_module.parameters()]) for model_module in model]),
+            ),
+            flush=True,
+        )
